+++ chartlets.js/packages/lib/src/plugins/vega/hooks/useSignalListeners.ts
@@ -5,11 +5,13 @@
  */
 
 import { useCallback, useMemo } from "react";
+import type { Result as VegaEmbedResult } from "vega-embed";
 import type { TopLevelSpec } from "vega-lite";
 
 import { type ComponentChangeHandler } from "@/index";
 import { isString } from "@/utils/isString";
 import { isObject } from "@/utils/isObject";
+import { useVegaSignalEmbed } from "./useVegaSignalEmbed";
 
 type SignalHandler = (signalName: string, signalValue: unknown) => void;
 
@@ -22,6 +24,11 @@ type SignalHandler = (signalName: string, signalValue: unknown) => void;
 type SelectionParameter = {
   name: string;
   select: "point" | "interval" | { type: "point" | "interval"; on: string };
+};
+
+type UseSignalListenersReturn = {
+  onEmbed: (result: VegaEmbedResult) => void;
+  signalListenerMap: Record<string, SignalHandler>;
 };
 
 const isSelectionParameter = (param: unknown): param is SelectionParameter =>
@@ -37,7 +44,7 @@ export function useSignalListeners(
   type: string,
   id: string | undefined,
   onChange: ComponentChangeHandler,
-): Record<string, SignalHandler> {
+): UseSignalListenersReturn {
   /*
    * Here, we create map of signals which will be then used to create the
    * map of signal-listeners because not all params are event-listeners, and we
@@ -65,7 +72,7 @@ export function useSignalListeners(
       }, signalNames);
   }, [chart]);
 
-  const handleClickSignal = useCallback(
+  const handleSignal = useCallback(
     (signalName: string, signalValue: unknown) => {
       if (id) {
         return onChange({
@@ -83,14 +90,14 @@ export function useSignalListeners(
    * Creates the map of signal listeners based on
    * the `signals` map computed above.
    */
-  return useMemo(() => {
+  const signalListenerMap = useMemo(() => {
     /*
      * Currently, we only have click events support, but if more are required,
      * they can be implemented and added in the map below.
      */
     const signalHandlers: Record<string, SignalHandler> = {
-      click: handleClickSignal,
-      drag: handleClickSignal,
+      click: handleSignal,
+      drag: handleSignal,
     };
 
     const signalListeners: Record<string, SignalHandler> = {};
@@ -104,5 +111,9 @@ export function useSignalListeners(
       }
     });
     return signalListeners;
-  }, [signalNames, handleClickSignal]);
+  }, [signalNames, handleSignal]);
+
+  const onEmbed = useVegaSignalEmbed(signalListenerMap);
+
+  return { onEmbed, signalListenerMap };
 }