+++ chartlets.js/packages/lib/vite.config.ts
@@ -78,6 +78,5 @@ export default defineConfig({
         return false;
       }
     },
-    exclude: ["**/vega/index.test.ts", "**/vega/VegaChart.test.tsx"],
   },
 });