+++ chartlets.js/CHANGES.md
@@ -2,6 +2,8 @@
 
 * Updated dependencies
 	- `glob: ^13.0.1`
+	- `react-vega: ^8.0.0`
+	- `vega-lite: ^6.4.1`
 	- `@vitest/coverage-istanbul: ^3.2.4`
 	- `vite: ^7.1.11`
 	- `vitest: ^3.2.4`
@@ -10,6 +12,10 @@
  (#124).
  
 * Added (MUI) component `Accordion`. (#41, #134)
+ (#124)
+ 
+* Adjusted `VegaChart` component, due to `react-vega` upgrade
+  from v7 to v8. (#132)
 
 * Fixed handling of `style` prop in `Tabs` component and added prop 
   `iconPosition`. (#135, #136)